Adding the rentBook(), returnBook() methods and Singleton implementation in BookService

+++ src/dao/UserDao.java
@@ -15,5 +15,5 @@ public interface UserDao {
 
     void delete(User user);
 
-    User read(int id);
+    User read(long id);
 }
